"Person Controller and postgress,mapper"

+++ src/main/java/com/crossasyst/jpademo/repository/PersonRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.RestController;
 
 
-public  interface PersonRepository extends JpaRepository<PersonEntity, Integer> {
+public  interface PersonRepository extends JpaRepository<PersonEntity, Long> {
 
 }
